+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add progressive previews in memory cache support to `FetchImage`
 - Add a property with `ImageContainer` to `FechImage`
 - `FetchImage` now supports optional request/publishesr (hanled as failrue)
+- Add an option to set default processors via `FetchImage`
 
 ## Nuke 10.1.0